Detach MenuLink hooks
